pyproject.toml
[project]
name = "gaidm-iwae"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch",
    "torchvision",
    "numpy",
    "matplotlib",
    "tqdm",
    "scipy",
    "pyyaml",
]
